refactor: relocate POST request for RFID
